Move literal parsing code into a separate file
Remove some dead code
